requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
redis==5.0.1
requests==2.31.0
pydantic==2.6.1
google-generativeai==0.3.2
python-multipart==0.0.9